+++ clean-code-katas/clean-code-katas.sln
@@ -29,6 +29,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fromromannumerals", "functi
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fizzbuzz.tests", "function-katas\tests\fizzbuzz.tests\fizzbuzz.tests.csproj", "{965020F0-3479-43AB-8EA7-79558A2CB2E1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fizzbuzz", "function-katas\fizzbuzz\fizzbuzz.csproj", "{AD5A1129-A66A-4F0E-9AAA-F814C6BED942}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -55,6 +57,10 @@ Global
 		{965020F0-3479-43AB-8EA7-79558A2CB2E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{965020F0-3479-43AB-8EA7-79558A2CB2E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{965020F0-3479-43AB-8EA7-79558A2CB2E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD5A1129-A66A-4F0E-9AAA-F814C6BED942}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD5A1129-A66A-4F0E-9AAA-F814C6BED942}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD5A1129-A66A-4F0E-9AAA-F814C6BED942}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD5A1129-A66A-4F0E-9AAA-F814C6BED942}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -66,6 +72,7 @@ Global
 		{04F1F976-3A5A-42C0-8B48-FEAA66138BB4} = {D06B579B-4270-40D4-AE72-481828C5C9DC}
 		{D6851572-558A-483B-90F6-8DA9AF03408B} = {A6B4F3FA-686E-44FC-B765-E35A7010999D}
 		{965020F0-3479-43AB-8EA7-79558A2CB2E1} = {D06B579B-4270-40D4-AE72-481828C5C9DC}
+		{AD5A1129-A66A-4F0E-9AAA-F814C6BED942} = {A6B4F3FA-686E-44FC-B765-E35A7010999D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {67C6EEA4-00F5-49D9-B1C1-68D845BC7AFD}
