+++ clean-code-katas/function-katas/tests/fizzbuzz.tests/FizzBuzzConverterTests.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using FluentAssertions;
+﻿using FluentAssertions;
 using Xunit;
 
 namespace fizzbuzz.tests
@@ -35,33 +32,6 @@ namespace fizzbuzz.tests
             var actual = _target.ConvertToFizzBuzzText(number);
 
             fizzBuzzText.Should().Be(actual);
-        }
-    }
-
-    internal class FizzBuzzConverter
-    {
-        internal string ConvertToFizzBuzzText(int number)
-        {
-            if (IsNumberDivisibleByThree(number) && IsNumberDivisibleByFive(number)) return "FizzBuzz";
-            if (IsNumberDivisibleByThree(number)) return "Fizz";
-            if (IsNumberDivisibleByFive(number)) return "Buzz";
-
-            return Convert.ToString(number);
-        }
-
-        private bool IsNumberDivisibleByThree(int number)
-        {
-            return number % 3 == 0;
-        }
-
-        private bool IsNumberDivisibleByFive(int number)
-        {
-            return number % 5 == 0;
-        }
-
-        public IEnumerable<string> ConvertToFizzBuzzTexts(IEnumerable<int> numbers)
-        {
-            return numbers.Select(n => ConvertToFizzBuzzText(n));
         }
     }
 }
